requirements.txt
africastalking==2.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==46.0.1
cssselect2==0.8.0
Deprecated==1.3.1
dj-database-url==3.0.1
Django==5.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.8.0
django-jazzmin==3.0.1
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.60.0
freetype-py==2.5.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.3.4
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
phonenumbers==9.0.13
pikepdf==10.1.0
pillow==12.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pycairo==1.28.0
pycparser==2.23
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pypdf==6.1.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
responses==0.25.8
rlPyCairo==0.4.0
schema==0.7.7
sendgrid==6.12.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.4
whitenoise==6.10.0
wrapt==2.0.1
xhtml2pdf==0.2.17
xlsxwriter==3.2.9
zopfli==0.2.3.post1
